fix(validation): allow numeric first line address
The existing address validation insists on at least 2 characters for an
address line to be valid.
Some users only put in their house number on the first line and the street
name on the second line. If the user lives in number 34 this is fine.
However if they live in number 9 their address will be rejected.

I have modified the regular expression for the first line of the address
to allow a single digit in that field.

Signed-off-by: 'Tom Wright' <thomas.wright2@dwp.gsi.gov.uk>