fix(postcode-validation): replace post code validator with a better one
The postcode validation regular expression has been replaced with the dwp one to ensure
compatibility and improved validation. The updated expressions is the same as the one
we use in java and in this public repo.
https://github.com/dwp/postcode-format-validation
The unit test have been extended to test a greater range of valid and invalid postcodes cases

Signed-off-by: Tom Wright <thomas.wright2@dwp.gsi.gov.uk>