fix: handle userland journey errors
Currently, when exceptions are thrown in fork or waypoint conditional
functions, they break the whole journey as it cannot be traversed
without that exception recurring each time. This fix catches those
exceptions and places a special "journey-fault" waypoint in the
journey at the point of failure. The user cannot move beyond this
waypoint until the exception is resolved.

Fixes #4

Signed-off-by: James Gauld <james.gauld@engineering.digital.dwp.gov.uk>