fix(gather-modifier): postcode reformatting accessibility changes
We found during user testing that some accessibility software
automatically adds spaces when entering letters and numbers, resulting in an invalid post code.
So the remedy is to remove all the spaces from the string and then to add one space back.
If this results in a valid postcode then great it returns the new postcode
If not then it returns the original value.
So if someone has made a mistake then they are not asked to correct the reformatted
text, but instead are asked to correct their original

Signed-off-by: Tom Wright <thomas.wright2@dwp.gsi.gov.uk>