refactor: whitelist fields with validators for session ingestion
Data that is to be ingested into the CASA session must now have an
associated field validator defined in the page's metadata.

BREAKING CHANGE: Where previously field validators were optional, they
are now mandatory if the data submitted for those fields is required to
be ingested into the CASA session.

Fixes #38