refactor(gather): check field validator condition prior to gathering
During gathering, only data that has a field validator is gathered into the
current session. We are currently only checking that a validator is
present for a field, and not checking that it is relevant in the current
journey data context. This change will check that the field validator
conditional returns true, and only then will include the field in the
gathered data set.