fix: traverse to correct page after edit if linear journey is shortened
For example if a page is removed from the end of a linear journey or a
page in the journey is no longer traversable:

                      Journey boundary
                      |
pregather  a--b--c--d-|-e <-- Edit origin
                      |
postgather a--b--c    |

In the existing code A, B and C will still match the pages in the
pregather so nextWaypoint will not be set to C, causing the user to end
up on E even though C requires action.

The changes mean that nextWaypoint will be set to the last page in the
postgather array if it is reached before the edit origin.

Signed-off-by: Niall Molloy <niall.molloy@engineering.digital.dwp.gov.uk>