refactor(validator): ensure waypointId is defined for queued pageMeta
Where the pageMeta object has not had a waypointId defined, the subsequent
validators may not recieve the correct field value (it will be undefined) and
thus not trigger correctly. Here we are ensuring that a waypoint id is present
in the meta object so data can be extracted from the JourneyContext correctly.