refactor: explicitly save session after regeneration to avoid race conditions
After calling endSession(), it is common to redirect the user to a subsequent page.
However, there is a potential race condition that can come into play if the
session store fails to save before the redirect and may cause undesired issues
with sessions unexpectedly expiring. Here we alter endSession() to wait for the
session to save before resolving the Promise.