perf: wrap main code to avoid validating options every on every call
If we warp the main code in another function and have that call itself 
we can avoid validating the input options on every loop. The main perf 
win though is that since the only thing that changes is the start 
position we can just pass that into the function each time, avoiding 
doing a clone of the options object every loop.