fix: route correctly when in sticky edit mode
In scenarios where a user has made a change, and the onward
journey is longer than the pre-modified journey, the user is
restricted to only proceeding as far as the last
waypoint in the pre-modified journey. This fix addresses
this behaviour, so that the user will be sent to the
furthest possible point in their modified journey.