fix: ensure validation state deserialises to ValidationError instances
Currently the validation state is deserialised from the session store
into a plain JS object. This works for the most part, but when attempting
to modify that state directly and writing back to the JourneyContext,
an exception is thrown as the context expects validation errors to be
stored as ValidationError instances.

Resolves #86