feat: pass session to context event handlers
On writing a context change to the session, this session is now made
available to any event listeners so that they may gather/manipulate
session data outside of just the current JourneyContext. This is
useful when you have multiple ephemeral contexts in the same session
and need to manipulate them based on events that are triggered on
a different context.