perf: serve static assets over non-paramterised paths
Presently, static assets are served from the router attached
to the - potentially - paramterised route on which a CASA app
is mounted. This means the browser may not be able to cache
assets if those URL parameters are unique. This change attaches
the static handlers to the non-paramterised portion of the
route. It is currently backwards compatible (serves assets
over both parameterised and non-parameterised routes), but
a future major update with remove support for the former.