refactor: reuse mountUrl as a means to simplify proxying setup
The backwards compatibility we introduced in ecdecded7c69e689821afc8c5c7092e966566296
that initially deprecated `mountUrl` in favour of `req.baseUrl`, has
actually been useful in discovering this alternative method of simplifying
the configuration of apps behind a path-rewriting proxy.

The `mountUrl` config option is now effectively used to indicate to
CASA that a proxy is in use and should execute its logic to maniuplate
`req.baseUrl`, stripping off the proxy prefix, for all subsequent
middleware.