fix: escape from sticky edit on return to edit origin in a proxied app
The `originalUrl` attribute is used by CASA to determine when it should
exit sticky-edit, using a simple comparison to the current request URL.

When running in proxy mode - where the express app is sitting behind a
forwarding proxy - the "originalUrl" attribute still contained the proxy
path prefix, and so the comparison failed, and the user was stuck in
edit mode.

This fix strips the proxy path prefix from `originalUrl` too.