feat: add JourneyContext.setSkipped() and JourneyContext.isSkipped()
`JourneyContext.setSkipped('page', { to: 'target'})` allows manually
setting a page as skipping to a specific target waypoint. This is now
used internally by the skip link logic.

`JourneyContext.isSkipped('page', { to: 'target'})` allows checking if a
page is skipping to a specific waypoint. This means you can now have a
page that skips to multiple waypoints, and check the target in your
plan:

```
plan.setRoute("options", "page-a", (r, c) => (
  c.isSkipped("options", { to: "page-a" })
));
plan.setRoute("options", "page-a", (r, c) => (
  c.isSkipped("options", { to: "page-b" })
));
```